Shuffle some tests around so they don't run in the integrationTest task
