+++ src/test/groovy/com/github/jrubygradle/storm/JRubyStormSpec.groovy
@@ -1,5 +1,6 @@
 package com.github.jrubygradle.storm
 
+import org.gradle.api.artifacts.Dependency
 import spock.lang.*
 
 import org.gradle.api.Project
@@ -104,5 +105,47 @@ class JRubyStormSpec extends Specification {
 
         then:
         project.configurations.findByName(JRubyStorm.DEFAULT_CONFIGURATION_NAME)
+    }
+
+    def "evaluation of the project should result in an assemble and run task"() {
+        given:
+        Project project = ProjectBuilder.builder().build()
+        project.apply plugin: 'com.github.jruby-gradle.storm'
+
+        when:
+        project.evaluate()
+
+        then:
+        project.tasks.findByName('assembleJRubyStorm')
+        project.tasks.findByName('runJRubyStorm')
+    }
+
+    def "evaluation of the project should result in dependencies being added to the configuration"() {
+        given:
+        Project project = ProjectBuilder.builder().build()
+        project.apply plugin: 'com.github.jruby-gradle.storm'
+        JRubyStorm task = project.task('spock', type: JRubyStorm)
+        def deps = task.configuration.dependencies
+
+        when:
+        project.evaluate()
+
+        then:
+        deps.matching { Dependency d -> d.name == 'redstorm' }
+    }
+
+    def "evaluation of the project should result in local mode dependencies"() {
+        given:
+        Project project = ProjectBuilder.builder().build()
+        project.apply plugin: 'com.github.jruby-gradle.storm'
+        JRubyStorm task = project.task('spock', type: JRubyStorm)
+
+        when:
+        project.evaluate()
+
+        then:
+        project.configurations.findByName('jrubyStormLocal')?.dependencies?.matching {
+            it.name == 'storm-core'
+        }
     }
 }
